Validate imported generic bounds
