+++ docs/COMPLETION_AUDIT.md
@@ -925,6 +925,9 @@ and do not assume Phase 4 is ready without evidence.
 - Imported public generic functions preserve behavior bounds whose behavior was
   imported by the source module, covered by
   `tests/zen/multi_file_imported_function_imported_behavior_bound/main.zen`.
+- Imported generic behavior bound arity diagnostics are covered through the
+  module graph by
+  `integration::imported_generic_behavior_bound_type_arg_arity_is_error`.
 - Imported public function parameter and return-type dependencies preserve
   behavior impl associations even when the entry module imports only the
   functions, covered by
